// src/domains/service-clients/order/proto/generated/order_service.ts
export interface GetRevenueStatsResponse {
  success?: RevenueStats | undefined;
  error?: Error | undefined;
}

export interface RevenueStats {
  revenueThisMonth: number;
  revenueLastMonth: number;
}

export interface GetRevenueStatsRequest {
  range: string;
}
// ...
export interface PlaceOrderRequest {
  userId: string;
  couponCode?: string | undefined;
  courseIds: string[];
}

export interface OrdersSuccess {
  orders: OrderData[];
  total: number;
}

export interface OrderSuccess {
  order: OrderData | undefined;
}

export interface OrderStatus {
  status: string;
  orderId: string;
}

export interface OrderResponse {
  success?: OrderSuccess | undefined;
  error?: Error | undefined;
}

export interface OrderStatusResponse {
  success?: OrderStatus | undefined;
  error?: Error | undefined;
}

export interface OrdersResponse {
  success?: OrdersSuccess | undefined;
  error?: Error | undefined;
}

export interface BookSessionRequest {
  userId: string;
  sessionId: string;
}

export interface GetOrderByIdRequest {
  orderId: string;
  userId: string;
}

export interface RestoreOrderRequest {
  orderId: string;
  userId: string;
}

export interface GetOrderStatusRequest {
  orderId: string;
}

export interface OrdersParams {
  page?: number | undefined;
  sortOrder?: string | undefined;
  pageSize?: number | undefined;
  status?: string | undefined;
}

export interface GetOrdersRequest {
  userId: string;
  params: OrdersParams | undefined;
}

export interface BookSessionResponse {
  success?: BookSessionSuccess | undefined;
  error?: Error | undefined;
}
// ...
export const BookSessionRequest: MessageFns<BookSessionRequest> = {
  encode(message: BookSessionRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userId !== "") {
      writer.uint32(10).string(message.userId);
    }
    if (message.sessionId !== "") {
      writer.uint32(18).string(message.sessionId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BookSessionRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBookSessionRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.sessionId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BookSessionRequest {
    return {
      userId: isSet(object.userId) ? globalThis.String(object.userId) : "",
      sessionId: isSet(object.sessionId) ? globalThis.String(object.sessionId) : "",
    };
  },

  toJSON(message: BookSessionRequest): unknown {
    const obj: any = {};
    if (message.userId !== "") {
      obj.userId = message.userId;
    }
    if (message.sessionId !== "") {
      obj.sessionId = message.sessionId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BookSessionRequest>, I>>(base?: I): BookSessionRequest {
    return BookSessionRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BookSessionRequest>, I>>(object: I): BookSessionRequest {
    const message = createBaseBookSessionRequest();
    message.userId = object.userId ?? "";
    message.sessionId = object.sessionId ?? "";
    return message;
  },
};

function createBaseGetOrderByIdRequest(): GetOrderByIdRequest {
  return { orderId: "", userId: "" };
}

export const GetOrderByIdRequest: MessageFns<GetOrderByIdRequest> = {
  encode(message: GetOrderByIdRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.orderId !== "") {
      writer.uint32(10).string(message.orderId);
    }
    if (message.userId !== "") {
      writer.uint32(18).string(message.userId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetOrderByIdRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetOrderByIdRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.orderId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetOrderByIdRequest {
    return {
      orderId: isSet(object.orderId) ? globalThis.String(object.orderId) : "",
      userId: isSet(object.userId) ? globalThis.String(object.userId) : "",
    };
  },

  toJSON(message: GetOrderByIdRequest): unknown {
    const obj: any = {};
    if (message.orderId !== "") {
      obj.orderId = message.orderId;
    }
    if (message.userId !== "") {
      obj.userId = message.userId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetOrderByIdRequest>, I>>(base?: I): GetOrderByIdRequest {
    return GetOrderByIdRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetOrderByIdRequest>, I>>(object: I): GetOrderByIdRequest {
    const message = createBaseGetOrderByIdRequest();
    message.orderId = object.orderId ?? "";
    message.userId = object.userId ?? "";
    return message;
  },
};
// ...
export const GetOrderStatusRequest: MessageFns<GetOrderStatusRequest> = {
  encode(message: GetOrderStatusRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.orderId !== "") {
      writer.uint32(10).string(message.orderId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetOrderStatusRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetOrderStatusRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.orderId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetOrderStatusRequest {
    return { orderId: isSet(object.orderId) ? globalThis.String(object.orderId) : "" };
  },

  toJSON(message: GetOrderStatusRequest): unknown {
    const obj: any = {};
    if (message.orderId !== "") {
      obj.orderId = message.orderId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetOrderStatusRequest>, I>>(base?: I): GetOrderStatusRequest {
    return GetOrderStatusRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetOrderStatusRequest>, I>>(object: I): GetOrderStatusRequest {
    const message = createBaseGetOrderStatusRequest();
    message.orderId = object.orderId ?? "";
    return message;
  },
};
// ...
export const GetOrdersRequest: MessageFns<GetOrdersRequest> = {
  encode(message: GetOrdersRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userId !== "") {
      writer.uint32(18).string(message.userId);
    }
    if (message.params !== undefined) {
      OrdersParams.encode(message.params, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetOrdersRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetOrdersRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.params = OrdersParams.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetOrdersRequest {
    return {
      userId: isSet(object.userId) ? globalThis.String(object.userId) : "",
      params: isSet(object.params) ? OrdersParams.fromJSON(object.params) : undefined,
    };
  },

  toJSON(message: GetOrdersRequest): unknown {
    const obj: any = {};
    if (message.userId !== "") {
      obj.userId = message.userId;
    }
    if (message.params !== undefined) {
      obj.params = OrdersParams.toJSON(message.params);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetOrdersRequest>, I>>(base?: I): GetOrdersRequest {
    return GetOrdersRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetOrdersRequest>, I>>(object: I): GetOrdersRequest {
    const message = createBaseGetOrdersRequest();
    message.userId = object.userId ?? "";
    message.params = (object.params !== undefined && object.params !== null)
      ? OrdersParams.fromPartial(object.params)
      : undefined;
    return message;
  },
};

function createBaseBookSessionResponse(): BookSessionResponse {
  return { success: undefined, error: undefined };
}

export const BookSessionResponse: MessageFns<BookSessionResponse> = {
  encode(message: BookSessionResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.success !== undefined) {
      BookSessionSuccess.encode(message.success, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BookSessionResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBookSessionResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.success = BookSessionSuccess.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BookSessionResponse {
    return {
      success: isSet(object.success) ? BookSessionSuccess.fromJSON(object.success) : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: BookSessionResponse): unknown {
    const obj: any = {};
    if (message.success !== undefined) {
      obj.success = BookSessionSuccess.toJSON(message.success);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BookSessionResponse>, I>>(base?: I): BookSessionResponse {
    return BookSessionResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BookSessionResponse>, I>>(object: I): BookSessionResponse {
    const message = createBaseBookSessionResponse();
    message.success = (object.success !== undefined && object.success !== null)
      ? BookSessionSuccess.fromPartial(object.success)
      : undefined;
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    return message;
  },
};
// ...
export const OrderServiceClient = makeGenericClientConstructor(
  OrderServiceService,
  "order_service.OrderService",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): OrderServiceClient;
  service: typeof OrderServiceService;
  serviceName: string;
};